refactor: merge conflict changes
