+++ packages/app/src/api/rights.ts
