+++ packages/docs/pages/devguide/new-component.mdx
@@ -43,7 +43,8 @@ function NewComponent(...) { ... }
 
 ---
 
-- import types with `import type { ... } from ...` syntax
+- import types with `import type { ... } from ...` synta
+  x
 
 ```typescript
 // ❌
@@ -86,7 +87,8 @@ function NewComponent(...) { ... }
 
 ---
 
-- export the functional component as default export (`export default MyComponent`) and at the bottom of the file
+- export the functional component as default export (`export default MyComponent`) and at the bottom of the fil
+  e
 
 ```tsx
 // ❌
