--- packages/lib/src/root/Document.tsx
@@ -1,44 +1,0 @@
-/*
- * Copyright (C) 2023 Frachtwerk GmbH, Leopoldstraße 7C, 76133 Karlsruhe.
- *
- * This file is part of Essencium Frontend.
- *
- * Essencium Frontend is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Essencium Frontend is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Essencium Frontend. If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { createGetInitialProps } from '@mantine/next'
-import _Document, { Head, Html, Main, NextScript } from 'next/document'
-
-const getInitialProps = createGetInitialProps()
-
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export class Document extends _Document {
-  static getInitialProps = getInitialProps
-
-  render(): JSX.Element {
-    return (
-      <Html lang="en">
-        <Head>
-          <link rel="icon" href="/img/web/favicon.ico" sizes="any" />
-          <link rel="apple-touch-icon" href="/img/web/apple-touch-icon.png" />
-        </Head>
-        <body>
-          <div id="notification" />
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-}
